style: componentiza alguns inputs do form
